Added .NET Standard 2.0 library
- Removed output files
- Removed VS files
